extract 3 core message processing methods to message_handler.rs
